requirements.txt
flask>=2.0.3
google-cloud-datastore==2.1.6
python-jose
flask-cors
six
python-dotenv>=0.19.2
requests>=2.27.1
authlib >= 1.0
json2html==1.3.0